AppData: Improve summary/description, OARS, indent
You don't have to go with this exact summary and description, but it's best to keep the summary as a descriptive tagline in sentence-case for consistency. And I cleaned up the description a little so it reads a little better in AppCenter without repeating itself as much.

Since your app uses an external web service, its `social-info` value **must** be set to `mild` or higher, depending on the data that is sent. You can always ensure your OARS data is correct by using the [official OARS generator](https://hughsie.github.io/oars/).

I also cleaned up the whitespace to use consistent indentation.